requirements.txt
selenium
requests
behave
allure_behave
Selenium-Screenshot
nose
pymysql2
PyMySQL
PyYaml
